backend-fastapi/knowledge-assistant-service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-openai
openai
pydantic-settings
